refactor: rename manage rebalance api modules
