+++ tests/e2e/demo/test_demo_scenarios.py
@@ -10,7 +10,7 @@ import pytest
 from fastapi.testclient import TestClient
 
 from src.api.main import app, get_db_session
-from src.api.routers.dpm_runs import reset_dpm_run_support_service_for_tests
+from src.api.routers.rebalance_runs import reset_dpm_run_support_service_for_tests
 from src.api.routers.proposals import reset_proposal_workflow_service_for_tests
 from src.core.common.capabilities import has_solver_dependencies
 from src.core.dpm.engine import run_simulation
